requirements.txt
arrow==1.3.0
dateparser==1.2.0
dearpygui==1.11.1
discord.py==2.3.2
Faker==25.5.0
gensim==4.3.2
googletrans==4.0.0rc1
matplotlib==3.9.0
mysql_connector_repackaged==0.3.1
networkx==3.3
nltk==3.8.1
num2words==0.5.13
pymorphy3==2.0.1
pymystem3==0.2.0
spacy==3.7.4
summa==1.2.0
text2num==2.5.1
vaderSentiment==3.3.2
vaderSentiment==3.3.2
psycopg2-binary
